Code cleanup, fixing DRY violation and fixing small frontend bug with ordering of command blocks.
Layout commands now define their logical order and the layout classes used
to build the corresponding layout structure.  The layout classes also define
whether they accept content.  This allows the logic regarding building
layouts to be simpler, and allows the frontend to order command blocks
without hard-coding special knowledge about the commands.

This fixes a bug with the ordering of 'inner row/col' with respect to 'new
row/col' in the editor.